my_site/requirements.txt
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
black==19.10b0
bottle==0.12.18
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.1
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.6
django-allauth==0.41.0
django-crispy-forms==1.8.1
django-rest-auth==0.9.5
djangorestframework==3.11.0
docopt==0.6.2
Flask==1.1.1
httpie==1.0.3
idna==2.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
path-and-address==2.0.1
pathspec==0.6.0
Pillow==6.2.1
Pygments==2.5.2
pylint==2.4.4
python3-openid==3.1.0
pytz==2019.3
regex==2019.12.20
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
sqlparse==0.3.0
toml==0.10.0
typed-ast==1.4.0
urllib3==1.25.7
waitress==1.4.0
Werkzeug==0.16.1
whitenoise==5.0.1
wrapt==1.11.2
